Always include `jsx` in list of Babylon plugins
JSX is in the project name, it should be explicit. Previously if custom
plugin(s) were specified then they had to also include `jsx`, to avoid
it being lost.

Update doc to reflect this. Remove `classProperties` from default list,
if required by a caller it should be specified as a command line arg
like any other plugin, it's not special as far as this repo is
concerned.